split tests for better clarity
